requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
boto3==1.41.1
botocore==1.41.1
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.121.3
h11==0.16.0
idna==3.11
jmespath==1.0.1
minio==7.2.18
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
s3transfer==0.15.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.38.0
